collab editor optimizations
