+++ src/components/views/CollaborativeNotebookView.tsx
@@ -3,17 +3,23 @@ import { CollaborativeEditor } from "../tiptap/tiptap-templates/collaborative/co
 import { FloatingToolbar } from "../tiptap/tiptap-templates/toolbar/FloatingToolbar";
 import { useNoteEditor } from "@/hooks/use-note-editor";
 import { useNoteContent } from "@/hooks/use-note-content";
-import { useRef, useState } from "react";
+import { useRef, useState, useMemo, memo } from "react";
 import { Clock } from "lucide-react";
 import { Editor } from "@tiptap/react";
 import { Loader2 } from "lucide-react";
 
-export const CollaborativeNotebookView = ({}) => {
+const CollaborativeNotebookView = memo(({}) => {
   const { currentNote, editorRef } = useNoteEditor();
   const { noteContent, isLoadingContent } = useNoteContent();
 
   const editorContainerRef = useRef<HTMLDivElement | null>(null);
   const [editor, setEditor] = useState<Editor | null>(null);
+
+  const lastSavedTime = useMemo(() => {
+    return currentNote
+      ? new Date(currentNote.updatedAt).toLocaleString()
+      : new Date().toLocaleString();
+  }, [currentNote]);
 
   return (
     <div className="relative min-h-screen bg-gradient-to-br from-background via-background to-background dark:from-background dark:via-background dark:to-background">
@@ -61,9 +67,7 @@ export const CollaborativeNotebookView = ({}) => {
 
               <div className="flex items-center gap-2">
                 <Clock className="h-3 w-3" />
-                <span>
-                  Last saved {new Date(currentNote.updatedAt).toLocaleString()}
-                </span>
+                <span>Last saved {lastSavedTime}</span>
               </div>
             </div>
           </div>
@@ -71,4 +75,8 @@ export const CollaborativeNotebookView = ({}) => {
       )}
     </div>
   );
-};
+});
+
+CollaborativeNotebookView.displayName = "CollaborativeNotebookView";
+
+export default CollaborativeNotebookView;
